dam-mcp/cloud_functions/meeting_notes_sync/requirements.txt
functions-framework==3.*
google-cloud-firestore>=2.0.0
google-api-python-client>=2.0.0
google-auth>=2.0.0
gspread>=6.0.0
